feat: 컴포넌트 분리 및 커스텀훅 추가
