+++ src/pages/Home/Home.module.scss
@@ -1,42 +1,5 @@
-@use "src/utils/styles/mediaQuery.scss" as media;
-
 .home {
   width: 100%;
-}
-
-.map-container {
-  display: none;
-
-  @include media.media-breakpoint-up(mobile) {
-    position: absolute;
-    right: 24px;
-    top: 24px;
-    width: 377px;
-    height: 57px;
-    border-radius: 100px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    font-size: 20px;
-    font-weight: 500;
-    line-height: 24.96px;
-    color: #222222;
-    z-index: 2;
-    backdrop-filter: blur(2px);
-    background-color: #ffffff99;
-    box-shadow: 1px 1px 10px 1px #0000001a;
-    cursor: pointer;
-  }
-
-  &__text {
-    display: flex;
-    margin-left: 24px;
-  }
-
-  &__image {
-    display: flex;
-    margin-right: 24px;
-  }
 }
 
 .map {
