+++ src/pages/Home/components/LocationInfo/LocationInfo.module.scss
@@ -1,0 +1,36 @@
+@use "src/utils/styles/mediaQuery.scss" as media;
+
+.map-container {
+  display: none;
+
+  @include media.media-breakpoint-up(mobile) {
+    position: absolute;
+    right: 24px;
+    top: 24px;
+    width: 377px;
+    height: 57px;
+    border-radius: 100px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    font-size: 20px;
+    font-weight: 500;
+    line-height: 24.96px;
+    color: #222222;
+    z-index: 2;
+    backdrop-filter: blur(2px);
+    background-color: #ffffff99;
+    box-shadow: 1px 1px 10px 1px #0000001a;
+    cursor: pointer;
+  }
+
+  &__text {
+    display: flex;
+    margin-left: 24px;
+  }
+
+  &__image {
+    display: flex;
+    margin-right: 24px;
+  }
+}
